// src/app/api/fn/sales/sales-payments-list.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { PaginatedCustomerPaymentList } from '../../models/paginated-customer-payment-list';

export interface SalesPaymentsList$Params {

/**
 * Customer id.
 */
  customer?: number;

/**
 * On or after this date.
 */
  date_from?: string;

/**
 * On or before this date.
 */
  date_to?: string;

/**
 * A page number within the paginated result set.
 */
  page?: number;

/**
 * Document status.
 */
  status?: 'POSTED' | 'VOID';
}

export function salesPaymentsList(http: HttpClient, rootUrl: string, params?: SalesPaymentsList$Params, context?: HttpContext): Observable<StrictHttpResponse<PaginatedCustomerPaymentList>> {
  const rb = new RequestBuilder(rootUrl, salesPaymentsList.PATH, 'get');
  if (params) {
    rb.query('customer', params.customer, {});
    rb.query('date_from', params.date_from, {});
    rb.query('date_to', params.date_to, {});
    rb.query('page', params.page, {});
    rb.query('status', params.status, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PaginatedCustomerPaymentList>;
    })
  );
}

salesPaymentsList.PATH = '/api/sales/payments/';